Cambio de nombres y Autorizacion 90%
