+++ app.js
@@ -8,8 +8,9 @@ var indexRouter = require('./routes/index');
 var usersRouter = require('./routes/users');
 var productosRouter = require('./routes/productos');
 var dashIndex = require('./routes/dash');
-var dashLogin = require('./routes/dash_login');
-var autorizar = require('./routes/dash_autorizar');
+var dashLogin = require('./routes/dashLogin');
+var dashPendientesAutorizar = require('./routes/dashPendientesAutorizar');
+
 
 var app = express();
 
@@ -26,9 +27,10 @@ app.use(express.static(path.join(__dirname, 'public')));
 app.use('/', indexRouter);
 app.use('/users', usersRouter);
 app.use('/productos', productosRouter);
-app.use('/dash_index', dashIndex);
-app.use('/dash_login', dashLogin);
-app.use('/autorizar', autorizar);
+app.use('/dashLogin', dashLogin);
+app.use('/dashIndex', dashIndex);
+app.use('/dashPendientesAutorizar', dashPendientesAutorizar);
+
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
