+++ views/dashIndex.ejs
@@ -1,0 +1,5 @@
+<%- include layouts/dashHeader.ejs %>
+  <div id="content" class="bg-grey w-100">
+  
+  </div>
+<%- include layouts/dashFooter.ejs %>
