+++ views/layouts/dashFooter.ejs
