+++ views/layouts/dashHeader.ejs
@@ -22,8 +22,12 @@
               <h4 class="text-light font-weight-bold mb-0">VideoGamesCity</h4>
           </div>
           <div class="menu">
-              <a href="#" class="d-block text-light p-3 border-0"><i class="fa fa-check-square" aria-hidden="true"></i>
+              <a href="/dashIndex" class="d-block text-light p-3 border-0"><i class="fa fa-home" aria-hidden="true"></i>
+                  Inicio</a>
+              <a href="/dashPendientesAutorizar" class="d-block text-light p-3 border-0"><i class="fa fa-check-square" aria-hidden="true"></i>
                   Autorizar Juegos</a>
+              <a href="#" class="d-block text-light p-3 border-0"><i class="fa fa-gamepad" aria-hidden="true"></i>
+                  Juegos</a>
               <a href="#" class="d-block text-light p-3 border-0"><i class="fa fa-user" aria-hidden="true"></i>
                   Perfil</a>
               <a href="#" class="d-block text-light p-3 border-0"><i class="fa fa-sign-out" aria-hidden="true"></i>
